Added Award 430HX with the Winbond W83877F Super I/O chip. Added Acer V35N (Award 430HX). Added Kaypro XT Clone and Achieve Systems 286. Added Award 430LX and Award 430NX. Added Intel Premiere/PCI II (Plato). Merged all stock PCem commits. Fixed a bug in the Intel 1 Mbit Flash BIOS emulation that caused the Premiere PCI/II to black screen. Hopefully final keyboard fix. Added the i450GX chipset for future usage. Added Intel 2 Mbit Flash BIOS emulation for future usage. Added Intel SIO (predecessor of PIIX) emulation for Award 430LX, Award 430NX, Intel Premiere/PCI and Intel Premiere/PCI II.

+++ PCem/cpu.h
@@ -55,7 +55,7 @@ typedef struct
         int cpu_type;
         int speed;
         int rspeed;
-        int multi;
+        double multi;
 	int pci_speed;
         uint32_t edx_reset;
         uint32_t cpuid_model;
@@ -89,7 +89,7 @@ extern CPU cpus_acer[];
 extern int cpu_iscyrix;
 extern int cpu_16bitbus;
 extern int cpu_busspeed;
-extern int cpu_multi;
+extern double cpu_multi;
 
 extern int cpu_hasrdtsc;
 extern int cpu_hasMSR;
